requirements.txt
Flask==2.2.3
numpy==1.24.2
tensorflow==2.11.1
opencv-python==4.7.0.72
pybase64==1.2.3
torch==2.0.0
pandas==2.0.0
Pillow==9.5.0
psutil==5.9.5
torchvision==0.15.1
PyYAML==6.0
tqdm==4.65.0
matplotlib==3.7.1
seaborn==0.12.2
ultralytics==8.0.84
Werkzeug==2.2.3
gunicorn==20.1.0

